+++ CHANGELOG.md
@@ -7,6 +7,7 @@ In Development
 Mail:
 
 * POP3S is now enabled (port 995).
+* In order to guard against misconfiguration that can lead to domain control validation hijacking, email addresses that begin with admin, administrator, postmaster, hostmaster, and webmaster can no longer be used for (new) mail user accounts, and aliases for these addresses may direct mail only to the box's administrator(s).
 
 System: